backend/requirements.txt
# NoirVision Backend Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.59.0
anthropic>=0.42.0
backboard-sdk>=1.5.0
boto3>=1.34.0
sqlmodel>=0.0.14
python-multipart>=0.0.6
